:truck: move "trust_registry_integration" test (really a verifier test) to the test_verifier module
